pyproject.toml
[tool.poetry]
name = "django-orm"
version = "0.1.0"
description = ""
authors = ["stevejeon <steve@mspw.io>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "*"
django-environ = "*"
psycopg2 = "*"
psycopg2-binary = "*"
django-extensions = "^3.1.3"
django-model-utils = "^4.1.1"
python-youtube = "^0.8.0"

[tool.poetry.dev-dependencies]
black = {version = "*", allow-prereleases = true}
ipdb = "*"
ipython = "*"
pre-commit = "*"
toml = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
